+++ packages/client/rollup.config.mjs
@@ -1,53 +1,51 @@
+import { external } from "@aminnairi/rollup-plugin-external";
+import commonjs from "@rollup/plugin-commonjs";
+import resolve from "@rollup/plugin-node-resolve";
+import typescript from "@rollup/plugin-typescript";
 import dts from "rollup-plugin-dts";
 
 /** @type {import('rollup').RollupOptions[]} */
 export default [
 	// ESM build
-	// {
-	// 	input: "src/index.ts",
-	// 	output: {
-	// 		file: "dist/index.js",
-	// 		format: "esm",
-	// 		sourcemap: true,
-	// 	},
-	// 	treeshake: false,
-	// 	external: [
-	// 		"@floating-ui/react",
-	// 		"@hookform/resolvers",
-	// 		"@lexical/react",
-	// 		"@lexical/utils",
-	// 		"@standard-schema/spec",
-	// 		"@tanstack/react-query",
-	// 		"@tanstack/react-query",
-	// 		"@tanstack/react-router",
-	// 		"@use-pico/common",
-	// 		"axios",
-	// 		"js-file-downloader",
-	// 		"kysely",
-	// 		"lexical",
-	// 		"localstorage-slim",
-	// 		"luxon",
-	// 		"react-dropzone",
-	// 		"react-hook-form",
-	// 		"react-hot-toast",
-	// 		"react",
-	// 		"read-excel-file",
-	// 		"sqlocal",
-	// 		"uuid",
-	// 		"write-excel-file",
-	// 		"zod",
-	// 		"zustand",
-	// 	],
-	// 	plugins: [
-	// 		resolve(),
-	// 		commonjs(),
-	// 		typescript({
-	// 			tsconfig: "./tsconfig.json",
-	// 			declaration: false,
-	// 			sourceMap: true,
-	// 		}),
-	// 	],
-	// },
+	{
+		input: "src/index.ts",
+		output: {
+			file: "dist/index.js",
+			format: "esm",
+			sourcemap: true,
+		},
+		treeshake: false,
+		external: [
+			//
+		],
+		plugins: [
+			resolve(),
+			commonjs(),
+			external({
+				include: [
+					"@hookform/resolvers/zod",
+					"@lexical/list",
+					"@lexical/react/LexicalAutoFocusPlugin",
+					"@lexical/react/LexicalComposer",
+					"@lexical/react/LexicalComposerContext",
+					"@lexical/react/LexicalContentEditable",
+					"@lexical/react/LexicalErrorBoundary",
+					"@lexical/react/LexicalHistoryPlugin",
+					"@lexical/react/LexicalListPlugin",
+					"@lexical/react/LexicalRichTextPlugin",
+					"@lexical/react/LexicalTabIndentationPlugin",
+					"react/jsx-runtime",
+					"sqlocal/kysely",
+				],
+			}),
+			typescript({
+				tsconfig: "./tsconfig.json",
+				declaration: false,
+				sourceMap: true,
+			}),
+		],
+	},
+
 	// DTS build
 	{
 		input: "src/index.ts",
@@ -59,7 +57,7 @@ export default [
 			dts(),
 		],
 		external: [
-			"@use-pico/common",
+			//
 		],
 	},
 ];