+++ packages/client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @use-pico/client
 
+## 4.0.49
+
+### Patch Changes
+
+- Fuck builds, pure source is still much better solution; keeping rollup here to show the pain I went through
+- Updated dependencies
+  - @use-pico/common@4.0.29
+
 ## 4.0.48
 
 ### Patch Changes